refactor(sdk): share tab wire mapping
